Styling of google button
